+++ src/game/renderables/models/PointLight.cpp
@@ -77,7 +77,7 @@ void PointLight::init()
 	}
 
 	// Store number of indices
-	indicesSize = indices.size();
+	indicesSize = static_cast<int>(indices.size());
 
 	glGenVertexArrays(1, &VAO);
 	glGenBuffers(1, &vertexVBO);
